# app/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenWatch (Stage 0)
  version: 0.1.0-alpha.1
  description: |
    Stage 0 walking-skeleton surface — three endpoints that exercise the
    foundation stack (correlation, idempotency, audit, error envelope).
servers:
  - url: https://localhost:8443
    description: Local development
paths:
...
  /api/v1/diagnostics:premium-echo:
    post:
      operationId: postDiagnosticsPremiumEcho
      summary: License-gated echo; requires premium_diagnostics feature
      x-required-feature: premium_diagnostics
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 255}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EchoRequest'}
      responses:
        '200':
          description: Echoed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EchoResponse'}
        '402':
          description: License does not include premium_diagnostics feature
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/admin/hosts:
    get:
      operationId: getAdminHosts
      summary: List hosts
      x-required-permission: host:read
      parameters:
        - {name: environment, in: query, required: false, schema: {type: string}}
        - {name: tag, in: query, required: false, schema: {type: string}}
      responses:
        '200':
          description: Host list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HostListResponse'}
        '403':
          description: Caller lacks host:read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    post:
      operationId: postAdminHosts
      summary: Create a host
      x-required-permission: host:write
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HostCreateRequest'}
      responses:
        '201':
          description: Host created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HostResponse'}
        '400':
          description: Validation failure (bad hostname, ip, port, etc.)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '409':
          description: Hostname already exists in environment
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/admin/hosts/{id}:
    get:
      operationId: getAdminHostByID
      summary: Fetch a host
      x-required-permission: host:read
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Host
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HostResponse'}
        '404':
          description: Host not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    patch:
      operationId: patchAdminHostByID
      summary: Update mutable host fields
      x-required-permission: host:write
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HostUpdateRequest'}
      responses:
        '200':
          description: Host updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HostResponse'}
        '400':
          description: Validation failure
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '404':
          description: Host not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    delete:
      operationId: deleteAdminHostByID
      summary: Soft-delete a host
      x-required-permission: host:delete
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '204':
          description: Deleted (deleted_at set)
        '404':
          description: Host not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
components:
...
  schemas:
...
    HostResponse:
      type: object
      required: [id, hostname, ip_address, port, environment]
      properties:
        id:           {type: string, format: uuid}
        hostname:     {type: string}
        ip_address:   {type: string}
        port:         {type: integer}
        display_name: {type: string}
        description:  {type: string}
        environment:  {type: string}
        tags:
          type: array
          items: {type: string}
        group_id:     {type: string, format: uuid, nullable: true}
        username:     {type: string}
        created_by:   {type: string, format: uuid}
        created_at:   {type: string, format: date-time}
        updated_at:   {type: string, format: date-time}
    HostListResponse:
      type: object
      required: [hosts]
      properties:
        hosts:
          type: array
          items: {$ref: '#/components/schemas/HostResponse'}
    HostCreateRequest:
      type: object
      required: [hostname, ip_address]
      properties:
        hostname:     {type: string, minLength: 1, maxLength: 256}
        ip_address:   {type: string, minLength: 1, maxLength: 64}
        port:         {type: integer, minimum: 1, maximum: 65535}
        display_name: {type: string, maxLength: 256}
        description:  {type: string, maxLength: 1024}
        environment:  {type: string, maxLength: 64}
        tags:
          type: array
          items: {type: string, maxLength: 64}
        group_id:     {type: string, format: uuid, nullable: true}
        username:     {type: string, maxLength: 256}
    HostUpdateRequest:
      type: object
      properties:
        ip_address:   {type: string, minLength: 1, maxLength: 64}
        port:         {type: integer, minimum: 1, maximum: 65535}
        display_name: {type: string, maxLength: 256}
        description:  {type: string, maxLength: 1024}
        environment:  {type: string, minLength: 1, maxLength: 64}
        tags:
          type: array
          items: {type: string, maxLength: 64}
        group_id:     {type: string, format: uuid, nullable: true}
        username:     {type: string, maxLength: 256}
...
    EchoRequest:
      type: object
      required: [message]
      properties:
        message: {type: string, minLength: 1, maxLength: 1024}
    EchoResponse:
      type: object
      required: [echoed, correlation_id]
      properties:
        echoed: {type: string}
        correlation_id: {type: string}
        audit_event_id: {type: string, format: uuid}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, fault, retryable, human_message]
          properties:
            code: {type: string}
            fault: {type: string, enum: [client, server, policy, external]}
            retryable: {type: boolean}
            human_message: {type: string}
            detail: {type: object, additionalProperties: true}
            correlation_id: {type: string}
...